Silence a useless Visual C 6 warning.
git-svn-id: svn://connect.creativelabs.com/trunk@790 109fce75-3012-0410-ac6a-932efec96c1b

+++ src/alutCodec.c
@@ -16,7 +16,7 @@ _alutCodecPCM8s (ALvoid *data, ALsizei length, ALint numChannels,
   ALsizei i;
   for (i = 0; i < length; i++)
     {
-      d[i] += 128;
+      d[i] += (int8_t)128;
     }
   return _alutBufferDataConstruct (data, length, numChannels, bitsPerSample,
                                    sampleFrequency);
